# docs/swagger.yaml
basePath: /
definitions:
  handler.UserResponse:
    properties:
      created_at:
        example: 04/09/2022
        type: string
      id:
        example: 1
        format: int64
        type: integer
      name:
        example: Admin
        type: string
      role:
        example: admin
        type: string
      username:
        example: admin
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a user api server.
  termsOfService: http://swagger.io/terms/
  title: GO User API
  version: "1.0"
paths:
...
  /api/users/{id}:
    get:
      consumes:
      - '*/*'
      description: Get User.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get User.
      tags:
      - Users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"